chore: Setting Interface 모듈 제거
