--- SimpleNote/App/CompositionRoot.swift
@@ -1,17 +1,0 @@
-//
-//  CompositionRoot.swift
-//  SimpleNote
-//
-//  Created by 홍성준 on 4/23/24.
-//
-
-import SettingFeatureImp
-import UIFeatureKit
-
-enum CompositionRoot {
-  
-  static func assemble() {
-    SettingAssemble.assemble()
-  }
-  
-}
