--- SimpleNote/Feature/Sources/SettingFeature/Setting/Bundle+.swift
@@ -1,14 +1,0 @@
-//
-//  Bundle+.swift
-//
-//
-//  Created by 홍성준 on 4/23/24.
-//
-
-import Foundation
-
-public extension Bundle {
-  
-  static let setting = Bundle.module
-  
-}
