--- SimpleNote/Feature/Sources/SettingFeatureImp/SettingAssemble.swift
@@ -1,17 +1,0 @@
-//
-//  SettingAssemble.swift
-//
-//
-//  Created by 홍성준 on 4/23/24.
-//
-
-import SettingFeature
-import UIFeatureKit
-
-public enum SettingAssemble {
-  
-  public static func assemble() {
-    SettingReducer.liveValue = .implement
-  }
-  
-}
